Support publicUrl in servers
