+++ packages/core/core/src/Parcel.js
@@ -1,11 +1,6 @@
 // @flow
 
-import type {
-  FilePath,
-  InitialParcelOptions,
-  ParcelOptions,
-  Stats
-} from '@parcel/types';
+import type {InitialParcelOptions, ParcelOptions, Stats} from '@parcel/types';
 import type {Bundle} from './types';
 import type InternalBundleGraph from './BundleGraph';
 
@@ -13,21 +8,15 @@ import AssetGraph from './AssetGraph';
 import {BundleGraph} from './public/BundleGraph';
 import BundlerRunner from './BundlerRunner';
 import WorkerFarm from '@parcel/workers';
-import TargetResolver from './TargetResolver';
-import getRootDir from '@parcel/utils/src/getRootDir';
 import nullthrows from 'nullthrows';
-import loadEnv from './loadEnv';
 import clone from 'clone';
-import path from 'path';
 import Cache from '@parcel/cache';
 import AssetGraphBuilder, {BuildAbortError} from './AssetGraphBuilder';
 import ConfigResolver from './ConfigResolver';
 import ReporterRunner from './ReporterRunner';
 import MainAssetGraph from './public/MainAssetGraph';
 import dumpGraphToGraphViz from './dumpGraphToGraphViz';
-
-// Default cache directory name
-const DEFAULT_CACHE_DIR = '.parcel-cache';
+import resolveOptions from './resolveOptions';
 
 export default class Parcel {
   initialOptions: InitialParcelOptions;
@@ -174,67 +163,6 @@ export default class Parcel {
     return Promise.all(promises);
   }
 }
-
-async function resolveOptions(
-  initialOptions: InitialParcelOptions
-): Promise<ParcelOptions> {
-  let entries: Array<FilePath>;
-  if (initialOptions.entries == null || initialOptions.entries === '') {
-    entries = [];
-  } else if (Array.isArray(initialOptions.entries)) {
-    entries = initialOptions.entries;
-  } else {
-    entries = [initialOptions.entries];
-  }
-
-  let rootDir =
-    initialOptions.rootDir != null
-      ? initialOptions.rootDir
-      : getRootDir(entries);
-
-  let targetResolver = new TargetResolver();
-  let resolvedTargets = await targetResolver.resolve(rootDir);
-
-  let targets;
-  if (initialOptions.targets) {
-    if (initialOptions.targets.length === 0) {
-      throw new Error('Targets was an empty array.');
-    }
-    targets = initialOptions.targets.map(target => {
-      if (typeof target === 'string') {
-        let matchingTarget = resolvedTargets.get(target);
-        if (!matchingTarget) {
-          throw new Error(`Could not find target with name ${target}`);
-        }
-        return matchingTarget;
-      }
-
-      return target;
-    });
-  } else {
-    targets = Array.from(resolvedTargets.values());
-  }
-
-  if (!initialOptions.env) {
-    await loadEnv(path.join(rootDir, 'index'));
-  }
-
-  let cacheDir =
-    initialOptions.cacheDir != null
-      ? initialOptions.cacheDir
-      : DEFAULT_CACHE_DIR;
-
-  // $FlowFixMe
-  return {
-    env: process.env,
-    ...initialOptions,
-    cacheDir,
-    entries,
-    rootDir,
-    targets
-  };
-}
-
 export {default as Asset} from './Asset';
 export {default as Dependency} from './Dependency';
 export {default as Environment} from './Environment';
