+++ packages/reporters/dev-server/src/ServerReporter.js
@@ -8,17 +8,15 @@ export default new Reporter({
     let serve = options.serve;
     if (!serve) return;
 
-    let isBrowser = options.targets.some(
-      target => target.env.context === 'browser'
-    );
-    if (!isBrowser) return;
+    let target = options.targets[0];
 
     let serverOptions = {
       ...serve,
       cacheDir: options.cacheDir,
-      distDir: 'dist', //options.distDir, // ! Not sure how this works now
-      // $FlowFixMe
-      publicUrl: options.publicUrl || '/'
+      distDir: target.distDir,
+      // Override the target's publicUrl as that is likely meant for production.
+      // This could be configurable in the future.
+      publicUrl: serve.publicUrl != null ? serve.publicUrl : '/'
     };
 
     let server = servers.get(serverOptions.port);
